New function to handle OM's taste to using what they call "matrix answers"
Basically they do this thing where they mix plaintext with base64 to dilimit seperate responses. They do this 1: I'm partly sure just to drive me bad 2: (more likely) they are sending via base64 as 'encryption' as is refernced in the documentation.

New function finds and filters these. So now any one with an 'opt-in' button will work. Which after testing turned out to be a LOT of them

phew